[counters] Move more implementation details to the counters module
- In general we're moving business logic out of the REST api
  blueprint code, to improve testability and general application
  structure.
